feat(precompiles): b20_asset V1 version seam (Phase 1, code motion)
Introduce the hardfork version seam on b20_asset and wrap today's behavior
as V1. Pure code motion — no logic change; existing b20 tests pass unedited.

- versions.rs: B20AssetVersion seam trait + B20AssetVersions::resolve(upgrade)
  -> Some(B20AssetVersionId::V1) from Beryl onward, else None.
- logic/{mod,v1}.rs: relocate the selector-routing layer (inner*, handle_b20_call,
  handle_asset_call, announce) verbatim; B20AssetV1 (ZST) implements the seam by
  delegating to the moved inner_with_observer routing.
- dispatch.rs: keep the shared entry (value/gas/init/observer/recorder); add
  dispatch_with_version which, after the init check, delegates to version.run.
  dispatch_with_observer is preserved as a V1 convenience for existing callers/tests.
- precompile.rs: create_precompile* resolve the version and return
  Option<DynPrecompile>; None (pre-Beryl) does not register. lookup.rs threads
  the None through.

Design note: B20AssetToken<S,P> is generic and dispatch threads a generic
observer O, so the seam's run<S,P,O> method is generic and the trait is
object-unsafe. Fork routing is therefore a small 'static Copy enum
(B20AssetVersionId) dispatched statically, not a &'static dyn trait object as
foo uses (foo's storage is concrete/'static; B20AssetStorage<'a> borrows ctx).
